restructure katana cli
